requirements.txt
# Backend Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Telegram Bot
python-telegram-bot==20.7

# Scheduling (для напоминаний)
apscheduler==3.10.4

# HTTP requests
httpx==0.25.2
aiohttp==3.9.1

# Payment (ЮKassa)
yookassa==3.0.0

# SMS
requests==2.31.0

# CORS
fastapi-cors==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.0
flake8==7.0.0
